refactor: 회원, 분석 관련 리팩토링 #24
