--- components/analytics/AnalyticsHeader.tsx
@@ -1,32 +1,0 @@
-// src/components/analytics/AnalyticsHeader.tsx
-import { Ionicons } from '@expo/vector-icons';
-import { router } from 'expo-router';
-import React from 'react';
-import { StyleSheet, TouchableOpacity } from 'react-native';
-import ScreenHeader from '@/components/shared/layout/ScreenHeader';
-import { Colors } from '@/constants/Colors';
-
-const AnalyticsHeader: React.FC = () => {
-    const handleProfilePress = () => {
-        router.push('/settings');
-    };
-
-    return (
-        <ScreenHeader
-            title="분석"
-            rightComponent={
-                <TouchableOpacity style={styles.profileIcon} onPress={handleProfilePress}>
-                    <Ionicons name="person-outline" size={24} color={Colors.text.secondary} />
-                </TouchableOpacity>
-            }
-        />
-    );
-};
-
-const styles = StyleSheet.create({
-    profileIcon: {
-        padding: 4,
-    },
-});
-
-export default AnalyticsHeader;
