+++ types/auth.ts
@@ -1,0 +1,12 @@
+export interface SignupFormData {
+    name: string;
+    email: string;
+    phoneNumber: string;
+    username: string;
+    password: string;
+    passwordConfirm: string;
+}
+
+export interface FieldErrors {
+    [key: string]: string;
+}
